// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/phrase-context-branches-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PhraseContextSidePageDto } from '../models/phrase-context-side-page-dto';
import { PhraseResolvedQueryDto } from '../models/phrase-resolved-query-dto';
import { PhraseSelectedPathDto } from '../models/phrase-selected-path-dto';
export interface PhraseContextBranchesResponse {
  activeBuildId: string;
  exactFullContextCount: number | null;
  following: PhraseContextSidePageDto;
  followingSelection: PhraseSelectedPathDto;
  previous: PhraseContextSidePageDto;
  previousSelection: PhraseSelectedPathDto;
  query: PhraseResolvedQueryDto;
  totalOccurrenceCount: number;
}